Add features to drive stats display
