+++ src/basestation/basestation_screen.cpp
@@ -2,8 +2,6 @@
 #include <basestation.hpp>
 
 #include <modules/console.hpp>
-#include <modules/network_config.hpp>
-#include <modules/drive_info.hpp>
 
 ScreenPositioning::ScreenPositioning(const nanogui::Vector2i& size, const nanogui::Vector2i& window_pos, int monitor, bool use_fullscreen) :
 	size(size),
@@ -20,12 +18,6 @@ BasestationScreen::BasestationScreen(const ScreenPositioning& pos)
 	perform_layout();
 	draw_all();
 	set_visible(true);
-
-	auto wnd = new NetworkConfig(this);
-	wnd->center();
-
-	new gui::DriveStats(this);
-
 }
 
 void BasestationScreen::set_fullscreen(GLFWmonitor* monitor) {
